2051 fixed
git-svn-id: https://subversion.assembla.com/svn/smart-cs/trunk@5656 37136b1e-4cb1-4992-8394-573b48a9bdce

+++ source/extensions/i2/org.wcs.smart.i2/src/org/wcs/smart/i2/handlers/ConservationAreaCloner.java
@@ -21,6 +21,7 @@
  */
 package org.wcs.smart.i2.handlers;
 
+import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -81,7 +82,7 @@ public class ConservationAreaCloner implements IConservationAreaTemplateCloner{
 			//clone record template
 			Path source = IntelReportManager.INSTANCE.getRecordTemplate(engine.getTemplateCa());
 			Path target = IntelReportManager.INSTANCE.getRecordTemplate(engine.getNewCa());
-			FileUtils.copyFile(source.toFile(), target.toFile());
+			if (Files.exists(source)) FileUtils.copyFile(source.toFile(), target.toFile());
 			monitor.worked(1);
 		}finally{
 			monitor.done();
